refactor: rename components on auth pages
