+++ frontend/src/app/sign-in/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { SignInForm } from "@/components/AuthForms/SignInForm/SignInForm";
+import { SignInForm } from "@/components/Auth/SignInForm/SignInForm";
 import { Container } from "@/components/Container/Container";
 
 export default function SignIn() {
