+++ frontend/src/app/sign-up/page.tsx
@@ -1,4 +1,4 @@
-import { SignUpForm } from "@/components/AuthForms/SignUpForm/SignUpForm";
+import { SignUpForm } from "@/components/Auth/SignUpForm/SignUpForm";
 import { Container } from "@/components/Container/Container";
 
 export default function SignUp() {
